Update about configuration.
