+++ src/main/resources/Help.txt
@@ -8,7 +8,7 @@
 §aMoneySAPI>>§fヘルプ
 §e/money see [対象者]  §f[対象者]の所持金を確認します。未入力だと自分の所持金が表示されます。
 §e/money pay [対象者] [金額]  §f[対象者]に[金額]を支払います。
-§e/money top  §f所持金ランキングのトップ３を表示します。（未実装）
+##§e/money top  §f所持金ランキングのトップ３を表示します。（未実装）
 §e/money debt  §fお金を借ります。（未実装）
 §e/money stock  §f株取引をします。（未実装）
 ::op
@@ -16,7 +16,7 @@
 §e/money take [対象者]　[金額]  §f[対象者]から[金額]を引きます。
 §e/money set [対象者]　[金額]  §f[対象者]の所持金を[金額]にします。
 ::all
-§d開発 - CosmoSunriseServerPluginEditorsTeam
-§f[Java版]popkechupki
+§d開発 - CoSSeDevelopmentTeam
+§f[Java版]popkechupki, itsu
 
-MoneySAPI Build2.0 Official version
+MoneySAPI Build4.0 Official version
